+++ manuals/sources/conf.py
@@ -28,7 +28,7 @@ author = u'Paulo Moura'
 # The short X.Y version
 version = u'3.25.0'
 # The full version, including alpha/beta/rc tags
-release = u'v3.25.0-b12'
+release = u'v3.25.0-b13'
 
 
 # -- General configuration ---------------------------------------------------